pathway_hackathon/requirements.txt
python-dotenv~=1.0
# Multi-agent support
crewai==0.28.8
langchain==0.1.20
langchain-openai==0.1.6
langchain-google-genai==1.0.3
streamlit==1.32.0
